+++ vera/__init__.py
@@ -1,4 +1,4 @@
 """Vera: a programming language designed for LLMs."""
 
-__version__ = "0.0.147"
-version = "0.0.147"
+__version__ = "0.0.148"
+version = "0.0.148"